example/clients/autogen/pyproject.toml
[project]
name = "autogen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.5.7",
    "autogen-ext[azure,mcp,openai]>=0.5.7",
    "pyyaml>=6.0.2",
]
